requirements.txt
numpy
pandas
scikit-learn
xgboost
lightgbm
matplotlib
seaborn
jupyter
notebook
pyyaml
joblib
dill
mlflow
fastapi
uvicorn
pydantic
python-dotenv
evidently
scipy
imbalanced-learn